test: extract common type and function into separated file
